unit test

+++ GoalTracker.Application/Goals/Commands/DeleteGoal/DeleteGoalCommandHandler.cs
@@ -3,6 +3,7 @@
 using AutoMapper;
 using GoalTracker.Application.Goals.Commands.CreateGoal;
 using GoalTracker.Application.Goals.Dtos;
+using GoalTracker.Domain;
 using GoalTracker.Domain.Entities;
 using GoalTracker.Domain.Exceptions;
 using GoalTracker.Domain.Repository;
@@ -11,7 +12,10 @@ using Microsoft.Extensions.Logging;
 
 namespace GoalTracker.Application.Goals.Commands.DeleteGoal;
 
-public class DeleteGoalCommandHandler(ILogger<DeleteGoalCommandHandler> logger,  IGoalsRepository goalsRepository) : IRequestHandler<DeleteGoalCommand>
+public class DeleteGoalCommandHandler(ILogger<DeleteGoalCommandHandler> logger,
+    IGoalsRepository goalsRepository,
+    IGoalAuthorizationService goalAuthorizationService
+    ) : IRequestHandler<DeleteGoalCommand>
 {
     public async Task Handle(DeleteGoalCommand request, CancellationToken cancellationToken)
     {
@@ -19,7 +23,8 @@ public class DeleteGoalCommandHandler(ILogger<DeleteGoalCommandHandler> logger, 
         var goal =  await goalsRepository.GetGoalAsync(request.Id);
         if (goal is null)
             throw new NotFoundException(nameof(Goal), request.Id.ToString());
-
+        if (!goalAuthorizationService.Authorize(goal, ResourceOperation.Delete))
+            throw new ForbidException();
         await goalsRepository.DeleteAsync(goal);
